requirements.txt
flask==2.0.1
Werkzeug==2.0.1
pdfplumber==0.7.4
python-docx==0.8.11
nltk==3.6.3
python-dotenv==0.19.1
